update fetch refresh liquidity & reward
