+++ src/app/containers/TradePage/index.tsx
@@ -1,9 +1,6 @@
 import React from 'react';
 
-import BigNumber from 'bignumber.js';
-import { BalancedJs } from 'packages/BalancedJs';
 import { useIconReact } from 'packages/icon-react';
-import { convertLoopToIcx } from 'packages/icon-react/utils';
 import { Helmet } from 'react-helmet-async';
 import { Flex, Box } from 'rebass/styled-components';
 
@@ -13,139 +10,24 @@ import LiquidityDetails from 'app/components/trade/LiquidityDetails';
 import LPPanel from 'app/components/trade/LPPanel';
 import ReturnICDSection from 'app/components/trade/ReturnICDSection';
 import SwapPanel from 'app/components/trade/SwapPanel';
-import bnJs from 'bnJs';
-import { useChangeLiquiditySupply } from 'store/liquidity/hooks';
+import { useFetchLiquidity } from 'store/liquidity/hooks';
 import { useFetchPrice } from 'store/ratio/hooks';
-import { useChangeReward } from 'store/reward/hooks';
+import { useFetchReward } from 'store/reward/hooks';
 import { useFetchBalance } from 'store/wallet/hooks';
 
 export function TradePage() {
-  // get ratio interval at here
-  useFetchPrice();
-
   const { account } = useIconReact();
+
+  useFetchPrice();
   useFetchBalance(account);
-  const changeLiquiditySupply = useChangeLiquiditySupply();
-  const changeReward = useChangeReward();
+  useFetchLiquidity(account);
+  useFetchReward(account);
 
   const [value, setValue] = React.useState<number>(0);
 
   const handleTabClick = (event: React.MouseEvent, value: number) => {
     setValue(value);
   };
-
-  const calculateTokenSupplied = (balance: BigNumber, poolTotal: BigNumber, totalSupply: BigNumber) => {
-    let tokenSupplied = balance
-      .multipliedBy(poolTotal)
-      .multipliedBy(new BigNumber(1).minus(balance.dividedBy(totalSupply)))
-      .dividedBy(totalSupply);
-    if (tokenSupplied.isNaN()) tokenSupplied = new BigNumber(0);
-    return tokenSupplied;
-  };
-
-  // get liquidity supply
-
-  const initLiquiditySupply = React.useCallback(() => {
-    if (account) {
-      Promise.all([
-        bnJs.Dex.getPoolTotal(BalancedJs.utils.sICXbnUSDpoolId.toString(), bnJs.sICX.address),
-        bnJs.Dex.getPoolTotal(BalancedJs.utils.sICXbnUSDpoolId.toString(), bnJs.bnUSD.address),
-        bnJs.Dex.balanceOf(BalancedJs.utils.sICXbnUSDpoolId.toString()),
-        bnJs.Dex.getTotalSupply(BalancedJs.utils.sICXbnUSDpoolId.toString()),
-
-        bnJs.Dex.getPoolTotal(BalancedJs.utils.BALNbnUSDpoolId.toString(), bnJs.Baln.address),
-        bnJs.Dex.getPoolTotal(BalancedJs.utils.BALNbnUSDpoolId.toString(), bnJs.bnUSD.address),
-        bnJs.Dex.balanceOf(BalancedJs.utils.BALNbnUSDpoolId.toString()),
-        bnJs.Dex.getTotalSupply(BalancedJs.utils.BALNbnUSDpoolId.toString()),
-
-        bnJs.Dex.getTotalSupply(BalancedJs.utils.sICXICXpoolId.toString()),
-        bnJs.eject({ account: account }).Dex.getICXBalance(),
-      ]).then(result => {
-        const [
-          sICXPoolsICXbnUSDTotal, // sm method `getPoolTotal`
-          bnUSDPoolsICXbnUSDTotal, // sm method `getPoolTotal`
-          sICXbnUSDBalance, // sm method `balanceOf`
-          sICXbnUSDTotalSupply, // sm method `totalSupply` pool sICXbnUSDpoolId
-
-          BALNPoolBALNbnUSDTotal,
-          bnUSDPoolBALNbnUSDTotal,
-          BALNbnUSDBalance,
-          BALNbnUSDTotalSupply,
-
-          sICXICXTotalSupply, // sm method `totalSupply` pool sICXICXpoolId
-          ICXBalance,
-        ] = result.map(v => convertLoopToIcx(v as BigNumber));
-        const sICXSuppliedPoolsICXbnUSD = calculateTokenSupplied(
-          sICXbnUSDBalance,
-          sICXPoolsICXbnUSDTotal,
-          sICXbnUSDTotalSupply,
-        );
-        const bnUSDSuppliedPoolsICXbnUSD = calculateTokenSupplied(
-          sICXbnUSDBalance,
-          bnUSDPoolsICXbnUSDTotal,
-          sICXbnUSDTotalSupply,
-        );
-
-        const BALNSuppliedPoolBALNbnUSD = calculateTokenSupplied(
-          BALNbnUSDBalance,
-          BALNPoolBALNbnUSDTotal,
-          BALNbnUSDTotalSupply,
-        );
-        const bnUSDSuppliedPoolBALNbnUSD = calculateTokenSupplied(
-          BALNbnUSDBalance,
-          bnUSDPoolBALNbnUSDTotal,
-          BALNbnUSDTotalSupply,
-        );
-
-        changeLiquiditySupply({
-          sICXPoolsICXbnUSDTotal,
-          bnUSDPoolsICXbnUSDTotal,
-          sICXbnUSDBalance,
-          sICXbnUSDTotalSupply,
-
-          BALNPoolBALNbnUSDTotal,
-          bnUSDPoolBALNbnUSDTotal,
-          BALNbnUSDBalance,
-          BALNbnUSDTotalSupply,
-
-          sICXSuppliedPoolsICXbnUSD,
-          bnUSDSuppliedPoolsICXbnUSD,
-
-          BALNSuppliedPoolBALNbnUSD,
-          bnUSDSuppliedPoolBALNbnUSD,
-
-          sICXICXTotalSupply,
-          ICXBalance,
-        });
-      });
-    }
-  }, [account, changeLiquiditySupply]);
-
-  React.useEffect(() => {
-    initLiquiditySupply();
-  }, [initLiquiditySupply]);
-
-  const initReward = React.useCallback(() => {
-    if (account) {
-      Promise.all([bnJs.Rewards.getRecipientsSplit(), bnJs.Rewards.getEmission(new BigNumber(1))]).then(result => {
-        const [poolsReward, poolEmission] = result.map(v => v);
-        const sICXICXreward = convertLoopToIcx(poolsReward['SICXICX']);
-        const sICXbnUSDreward = convertLoopToIcx(poolsReward['SICXbnUSD']);
-        const BALNbnUSDreward = convertLoopToIcx(poolsReward['BALNbnUSD']);
-        const poolDailyReward = convertLoopToIcx(poolEmission);
-        changeReward({
-          sICXICXreward,
-          sICXbnUSDreward,
-          BALNbnUSDreward,
-          poolDailyReward,
-        });
-      });
-    }
-  }, [account, changeReward]);
-
-  React.useEffect(() => {
-    initReward();
-  }, [initReward]);
 
   // update the width on a window resize
   const ref = React.useRef<HTMLDivElement>();
